コード修正・整理
ファイル名・クラス名を変更
AppBarの追加
