+++ lib/schedule/components/build_date_label.dart
@@ -5,7 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:table_calendar/table_calendar.dart';
 
 import 'package:flutter_tutorial/constants.dart';
-import 'package:flutter_tutorial/calendar/view_model/calendar_state_providers.dart';
+import 'package:flutter_tutorial/schedule/view_model/calendar_state_providers.dart';
 
 class BuildDateLabel extends ConsumerWidget {
   const BuildDateLabel({super.key});
