+++ lib/schedule/model/schedule_model.dart
