+++ lib/schedule/model/schedule_model.freezed.dart
