+++ lib/schedule/view_model/calendar_state_providers.dart
